requirements.txt
datapizza-ai>=0.0.9
aiohttp>=3.9.0
httpx>=0.25.0
scipy>=1.11.0
numpy>=1.24.0
python-dotenv>=1.0.0
